add namedatomicstatus to zaino-common
(cherry picked from commit 1de35c598da6d2ac1d1213d2038593335dfaa817)
