+++ packages/zaino-state/src/lib.rs
@@ -117,7 +117,7 @@ pub use error::{LegacyRpcError, NodeBackedIndexerServiceError};
 
 pub(crate) mod status;
 
-pub use status::{AtomicStatus, NamedAtomicStatus, Status, StatusType};
+pub use status::{NamedAtomicStatus, Status, StatusType};
 
 pub(crate) mod stream;
